Commit of #148
The reading of an input port have been changed to allow higher reading intervals. The last poll time of the input port is stored in the gpio_pin, an interval has been added to the input pin.
This removes the need for the 10mSec wait between two read intervals (which was previously used) and results in unusable results if there are more PCF8574 devices in the configuration (see Issue #148 )